% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_variables.R
\name{count_variables}
\alias{count_variables}
\title{count the number of times each variable is used in a ranger random forest}
\usage{
count_variables(object)
}
\arguments{
\item{object}{a ranger forest object}
}
\description{
count the number of times each variable is used in a ranger random forest.
help(treeInfo) warns
"splitvarID -- ID of the splitting variable, 0-indexed. Caution, the variable order changes if the formula interface is used"
this should be investigated
}
\examples{
\dontrun{
library(ranger)
rf1<-ranger(Species ~ ., data = iris,importance="impurity",seed=123)
count_variables(rf1)
}
}
\keyword{counts}
