% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/determine.C.R
\name{determine.C}
\alias{determine.C}
\title{determine.C}
\usage{
determine.C(f_fit)
}
\arguments{
\item{f_fit}{Do you love cats? Defaults to TRUE.}
}
\description{
This function allows you to express your love of cats.
}
\examples{
cat_function()
}
\keyword{cats}
