% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/determine_cutoff.R
\name{determine_cutoff}
\alias{determine_cutoff}
\title{evaluate a measure that can be used to determing the varaible and cutoff value for a RF model}
\usage{
determine_cutoff(
  imp,
  t2,
  cutoff = c(0, 1, 4, 10, 15, 20),
  Q = 0.75,
  plot = NULL
)
}
\arguments{
\item{imp}{vector of MDI varaible importances importances}

\item{t2}{number of times each variable is used in the a ranger forest. Returned by count_variables for a ranger RF}

\item{cutoff}{values to evaluate}

\item{Q}{}

\item{plot}{}
}
\description{
evaluate a measure that can be used to determing the varaible and cutoff value for a RF model
}
\examples{
library(ranger)
rf1 <-ranger(Species ~ ., data = iris,importance="impurity",seed=123)
imp<-log(importance(rf1))
t2 <- count_variables(rf1)
aa <- determine_cutoff(imp, t2, cutoff=c(0,1,2), Q = 0.75, plot = NULL)
}
