% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f.spline.R
\name{f.spline}
\alias{f.spline}
\title{fit a spline}
\usage{
f.spline(zz, df = 10)
}
\arguments{
\item{zz}{the varaible importances}

\item{df}{the degrees of freedom for the spline fit}
}
\description{
fit a spline
}
\keyword{spline}
