% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imp1.R
\docType{data}
\name{imp1}
\alias{imp1}
\title{2000 importance values}
\format{
A vector with 2000 values
\describe{
\item{imp1}{importaces }
}
}
\source{
\url{http://www.diamondse.info/}
}
\usage{
imp1
}
\description{
A dataset containing  2000 importance values
}
\keyword{datasets}
