% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_PIMP.R
\name{my_PIMP}
\alias{my_PIMP}
\title{my_PIMP
based on the PIMP function from the vita package. ‘PIMP’ implements the test approach of Altmann et al. (2010) for
the permutation variable importance measure ‘VarImp’ returned by the randomForest package (Liaw and Wiener (2002)) for
classification and regression.}
\usage{
my_PIMP(X, y, rForest, S = 100, parallel = FALSE, ncores = 0, seed = 123, ...)
}
\arguments{
\item{X}{}

\item{y}{}

\item{rForest}{}

\item{S}{}

\item{parallel}{}

\item{ncores}{}

\item{seed}{}
}
\value{
an object of class PIMP
}
\description{
my_PIMP applies the same method as PIMP but to the MDI (mean decrease in impurity)  variable importance
(mean decrease in Gini index for classification and  mean decrease in MSE for regression).
}
\examples{
library(RFlocalfdr)
library(vita) #vita: Variable Importance Testing Approaches
data(smoking)
?smoking 
y<-smoking$y
y<-factor(y)
smoking_data<-smoking$rma

cl.ranger <- ranger(y=y, x=smoking_data,mtry = 3,num.trees = 1000, importance = 'impurity')
system.time(pimp.varImp.cl<-my_ranger_PIMP(X,y,cl.ranger,S=10, parallel=TRUE, ncores=3)) 
pimp.t.cl <- PimpTest(pimp.varImp.cl,para = FALSE)
aa <- summary(pimp.t.cl,pless = 0.05)
length(which(aa$cmat2[,"p-value"]< 0.05))
hist(aa$cmat2[,"p-value"],breaks=20)
}
