% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotQ.R
\name{plotQ}
\alias{plotQ}
\title{plotQ}
\usage{
plotQ(imp, debug.flag = 0, temp.dir = NULL, try.counter = 3)
}
\arguments{
\item{imp}{"reduction in impurity" importances from a random forest model}

\item{debug.flag}{either 0 (no debugging information), 1 or 2}

\item{temp.dir}{if debug flag is >0 then information is written to temp.dir}

\item{try.counter}{where to explain this?}
}
\value{
\itemize{
\item df
\item final.estimates_C_0.95
\item final.estimates_cc
\item temp.dir
\item C_0.95
\item cc
\item fileConn
\item f_fit
\item ww
\item aa_cc
}
}
\description{
produces a plot showing the q values
\itemize{
\item q_95, the 95th quantile of the data
\item q using the penalized selection method of Gauran et.al 2018
}
}
\details{
We estiamte a value "q" such that:
to the left of "q", the density is somposed solely of NULL importance values
to the right of "q" we have  a density that is a mixture of null and non-null  importance values.
The method of Gauran et.al 2018 may not work in cases where the data distribution is not well modelled by a skew-normal.
The q_95 value can be uses as a workaround in these case.
In many cases they will be very similar
}
\examples{
data(ch22)                                                                                 
?ch22                                                                                     
#document how the data set is created                                                      
plot(density(log(ch22$imp)))                                                               
t2 <-ch22$C                                                                                
imp<-log(ch22$imp)                                                                         
#Detemine a cutoff to get a unimodal density.
# This was calculated previously. See determine_cutoff
imp<-imp[t2 > 30]
qq <- plotQ(imp,debug.flag = 0)

data(smoking)
?smoking 
y<-smoking$y
smoking_data<-smoking$rma
y.numeric <-ifelse((y=="never-smoked"),0,1)

library(ranger)
rf1 <-ranger(y=y.numeric ,x=smoking_data,importance="impurity",seed=123, num.trees = 10000,
             classification=TRUE)
t2 <-count_variables(rf1)
imp<-log(rf1$variable.importance)
#png("./supp_figures/smoking_log_importances.png")
plot(density(imp),xlab="log importances",main="")
cutoffs <- c(2,3,4,5)
#png("./supp_figures/smoking_data_determine_cutoff.png")
res.con<- determine_cutoff(imp,t2,cutoff=cutoffs,plot=c(2,3,4,5))
#dev.off()

#png("./supp_figures/smoking_data_determine_cutoffs_2.png")
plot(cutoffs,res.con[,3],pch=15,col="red",cex=1.5,ylab="max(abs(y - t1))")
#dev.off()
cutoffs[which.min(res.con[,3])]

temp<-imp[t2 > 3]
temp <- temp - min(temp) + .Machine$double.eps
qq <- plotQ(temp,debug.flag = 1)
ppp<-run.it.importances(qq,temp,debug.flag = 0)
aa<-significant.genes(ppp,temp,cutoff=0.05,debug.flag=0,do.plot=TRUE,use_95_q=TRUE)
length(aa$probabilities) # 17
}
\keyword{importance}
\keyword{variable}
