% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propTrueNullByLocalFDR.R
\name{propTrueNullByLocalFDR}
\alias{propTrueNullByLocalFDR}
\title{propTrueNullByLocalFDR}
\usage{
propTrueNullByLocalFDR(p)
}
\arguments{
\item{p}{probabilities}
}
\description{
propTrueNullByLocalFDR
}
\examples{
propTrueNullByLocalFDR()
}
\keyword{cats}
