% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{smoking}
\alias{smoking}
\title{Effects of cigarette smoke on the human airway epithelial cell transcriptome}
\format{
A list with rma (the transcript data) and y (the class labels):
\describe{
\item{rma}{ 51 by  22283, log2 real values }
\item{y}{a character vector, "smoking" and "never-smoked"}
...
}
}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE994}
}
\usage{
smoking
}
\description{
A dataset containing normalized transcript measurements for 51 subjects and 22283 transcripts.
See Spira et al (2004). "Gene Expression Profiling of Human Lung Tissue from Smokers with Severe Emphysema",
Am J Respir Cell Mol Biol.
}
\keyword{datasets}
