% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/determine_cutoff.R
\name{determine_cutoff}
\alias{determine_cutoff}
\title{count the number of times each variable is used in the a ranger forest}
\usage{
determine_cutoff(
  imp,
  t2,
  cutoff = c(0, 1, 4, 10, 15, 20),
  Q = 0.75,
  plot = NULL
)
}
\arguments{
\item{imp}{importances}

\item{t2}{number of times each variable is used in the a ranger forest.}

\item{cutoff}{}

\item{Q}{}

\item{plot}{}
}
\description{
count the number of times each varaible is used in the a ranger forest.
help(treeInfo) warns
"splitvarID -- ID of the splitting variable, 0-indexed. Caution, the variable order changes if the formula interface is used"
this should be investigated
}
\examples{
library(ranger)
rf1 <-ranger(Species ~ ., data = iris,importance="impurity",seed=123)
imp<-log(importance(rf1))
t2 <- count_variables(rf1)
aa <- determine_cutoff(imp, t2, cutoff=c(0,1,2), Q = 0.75, plot = NULL)
}
