% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f.fit.R
\name{f.fit}
\alias{f.fit}
\title{fit a spline to the histogram of zz}
\usage{
f.fit(zz, df = 10, debug.flag = 0, temp.dir = NULL)
}
\arguments{
\item{zz}{the variable importances}

\item{df}{the degrees of freedom for the spline fit}
}
\value{
a list with the following components
"x" -- midpoints of the histogram
"zh" -- a histogram object as returned by "hist"
"f.spline" -- the spline fit. The fit is given by a glm mode glm(zh$counts ~ splines:::ns(x), poisson)
"counts" the counts from the histogram
}
\description{
fit a spline to the histogram of zz
}
\examples{
data(imp1)                          
res <- f.fit(imp1)                  
plot(res$zh)                        
plot(res$x,res$counts)              
plot(res$zh$breaks[-1],res$f.spline)
}
\keyword{spline}
