% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f.fit.R
\name{f.fit}
\alias{f.fit}
\title{fit a spline}
\usage{
f.fit(zz, df = 10, debug.flag = 0)
}
\arguments{
\item{zz}{the varaible importances}

\item{df}{the degrees of freedom for the spline fit}
}
\description{
fit a spline
}
\keyword{spline}
