% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.to.data.set.R
\name{fit.to.data.set}
\alias{fit.to.data.set}
\title{fit.to.data.set}
\usage{
fit.to.data.set(
  df,
  imp,
  debug.flag = 0,
  plot.string = "",
  temp.dir = NULL,
  try.counter = 3,
  return.all = FALSE
)
}
\arguments{
\item{imp}{importances}

\item{debug.flag}{debug flag
#' @keywords cats}

\item{f_fit}{object returned by f.fit}
}
\description{
This function fit a skew normal to a set of data
}
\examples{
data(ch22)                                       
imp<-log(ch22$imp)                               
t2<-ch22$C                                       
temp<-imp[t2 > 30]   #                           
temp<-temp[temp != -Inf]                         
temp <- temp - min(temp) + .Machine$double.eps   
f_fit <- f.fit(temp)                             
y <- f_fit$zh$density                            
x <- f_fit$midpoints                             
df <- data.frame(x, y)                           
mm.df3 <- fit.to.data.set(df, temp)              
mm.df3
##              Estimate Std..Error  t.value     Pr...t..
## xi.xi        1.102303 0.03669284 30.04136 1.485263e-56
## omega.omega  1.246756 0.04716184 26.43569 6.276349e-51
## lambda.alpha 1.799169 0.17343872 10.37351 3.103195e-18
}
