% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.to.data.set.R
\name{fit.to.data.set}
\alias{fit.to.data.set}
\title{fit.to.data.set}
\usage{
fit.to.data.set(
  df,
  imp,
  debug.flag = 0,
  plot.string = "",
  temp.dir = NULL,
  try.counter = 3,
  return.all = FALSE
)
}
\arguments{
\item{imp}{importances}

\item{debug.flag}{debug flag
#' @keywords cats}

\item{f_fit}{object returned by f.fit}
}
\description{
This function allows you to express your love of cats.
}
\examples{
cat_function()
}
