% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.to.data.set.wrapper.R
\name{fit.to.data.set.wrapper}
\alias{fit.to.data.set.wrapper}
\title{fit.to.data.set.wrapper}
\usage{
fit.to.data.set.wrapper(
  df,
  imp,
  debug.flag = 0,
  plot.string = "",
  temp.dir = temp.dir,
  try.counter = 3
)
}
\arguments{
\item{df, }{contains x and y, midpoints and counts from a histogram of imp}

\item{imp}{importances}

\item{debug.flag}{debug flag}

\item{plot.string, }{file name for a debugging plot}

\item{temp.dir, }{directory for debugging output}

\item{try.counter}{try.counter=1 my.dsn xi=  1
try.counter=2  xi=  mean(x)
try.counter=3 start xi, omega, lambda from the parameters retuned by fitdistrplus::fitdist}
}
\description{
This function allows you to express your love of cats.
}
\examples{
cat_function()
}
