% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local.fdr.R
\name{local.fdr}
\alias{local.fdr}
\title{local fdr}
\usage{
local.fdr(
  f,
  x,
  FUN = dburr,
  p0 = 1,
  debug.flag = 0,
  plot.string = "",
  temp.dir = NULL,
  ...
)
}
\arguments{
\item{f}{f}

\item{x}{x}

\item{FUN}{FUN}
}
\description{
calculate the local
}
\examples{
local.fdr()
}
\keyword{cats}
