% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local.fdr.R
\name{local.fdr}
\alias{local.fdr}
\title{local fdr}
\usage{
local.fdr(f, x, FUN = dburr, p0 = 1, ...)
}
\arguments{
\item{f}{f}

\item{x}{x}

\item{FUN}{FUN}
}
\description{
This function allows you to express your love of cats.
}
\examples{
local.fdr()
}
\keyword{cats}
