% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my.dsn.R
\name{my.dsn}
\alias{my.dsn}
\title{density of skew-normal}
\usage{
my.dsn(x, xi = 0, omega = 1, lambda = 1)
}
\arguments{
\item{x}{param}

\item{xi}{param}

\item{omega}{param}

\item{lambda}{param}
}
\description{
This function allows you to express your love of cats.
}
\keyword{cats}
