% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propTrueNullByLocalFDR.R
\name{propTrueNullByLocalFDR}
\alias{propTrueNullByLocalFDR}
\title{propTrueNullByLocalFDR}
\usage{
propTrueNullByLocalFDR(p)
}
\arguments{
\item{p}{probabilities}
}
\value{
An estimate of the proportion of null p-values by the local fdr
}
\description{
Estimate proportion of NULL p-values. Based on  .propTrueNullByLocalFDR in  limma: Linear Models for Microarray Data
written by Belinda Phipson and Gordon Smyth
}
\keyword{cats}
