% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.it.importances.R
\name{run.it.importances}
\alias{run.it.importances}
\title{run.it.importances}
\usage{
run.it.importances(imp, debug.flag = 0, temp.dir = NULL, try.counter = 3)
}
\arguments{
\item{imp}{"reduction in impurity" importances from a random forest model}

\item{debug.flag}{either 0 (no debugging information), 1 or 2}

\item{temp.dir}{if debug flag is >0 then information is written to temp.dir}

\item{try.counter}{where to explain this?}
}
\description{
run.it.importances
}
\examples{
data(ch22)                                                                                 
? ch22                                                                                     
#document how the data set is created                                                      
plot(density(log(ch22$imp)))                                                               
t2 <-ch22$C                                                                                
imp<-log(ch22$imp)                                                                         
#Detemine a cutoff to get a unimodal density.                                              
res.temp <- determine_cutoff(imp, t2 ,cutoff=c(1,10,20,30),plot=c(1,10,20,30),Q=0.75)      
plot(c(1,2,3,4),res.temp[,3])                                                              
                                                                                           
res.temp <- determine_cutoff(imp, t2 ,cutoff=c(25,30,35,40),plot=c(25,30,35,40),Q=0.75)    
plot(c(25,30,35,40),res.temp[,3])                                                          
imp<-imp[t2 > 30]                                                                          
ppp<-run.it.importances(imp,debug=0)                                                       
aa<-significant.genes(ppp,imp,cutoff=0.2,debug.flag=0,do.plot=2)                           
length(aa$probabilities) # 6650                                                            
aa<-significant.genes(ppp,imp,cutoff=0.05,debug.flag=0,do.plot=2)                          
length(aa$probabilities) # 3653                                                            
}
\keyword{importance}
\keyword{variable}
