% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.it.importances.R
\name{run.it.importances}
\alias{run.it.importances}
\title{run.it.importances}
\usage{
run.it.importances(imp1, debug.flag = 0)
}
\description{
This function allows you to express your love of cats.
#' @param imp importances
#' @param debug.flag debug
#' @keywords cats
}
\examples{
cat_function()
}
