% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/significant.genes.R
\name{significant.genes}
\alias{significant.genes}
\title{significant.genes}
\usage{
significant.genes(object, imp, cutoff = 0.2, do.plot = 0, Debug = 0)
}
\arguments{
\item{object}{object}

\item{imp}{importances}

\item{cutoff}{cutoff}

\item{do.plot}{do.plot}

\item{Debug}{Debug}
}
\description{
This function allows you to express your love of cats.
}
\examples{
cat_function()
}
\keyword{cats}
